+++ packages/location/example/lib/listen_location.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:location/location.dart';
 
 class ListenLocationWidget extends StatefulWidget {
@@ -14,16 +15,21 @@ class _ListenLocationState extends State<ListenLocationWidget> {
   final Location location = Location();
 
   LocationData? _location;
-  late StreamSubscription<LocationData> _locationSubscription;
+  StreamSubscription<LocationData>? _locationSubscription;
   String? _error;
 
   Future<void> _listenLocation() async {
     _locationSubscription =
         location.onLocationChanged.handleError((dynamic err) {
+      if (err is PlatformException) {
+        setState(() {
+          _error = err.code;
+        });
+      }
+      _locationSubscription?.cancel();
       setState(() {
-        _error = err.code;
+        _locationSubscription = null;
       });
-      _locationSubscription.cancel();
     }).listen((LocationData currentLocation) {
       setState(() {
         _error = null;
@@ -31,10 +37,23 @@ class _ListenLocationState extends State<ListenLocationWidget> {
         _location = currentLocation;
       });
     });
+    setState(() {});
   }
 
   Future<void> _stopListen() async {
-    _locationSubscription.cancel();
+    _locationSubscription?.cancel();
+    setState(() {
+      _locationSubscription = null;
+    });
+  }
+
+  @override
+  void dispose() {
+    _locationSubscription?.cancel();
+    setState(() {
+      _locationSubscription = null;
+    });
+    super.dispose();
   }
 
   @override
@@ -52,12 +71,13 @@ class _ListenLocationState extends State<ListenLocationWidget> {
               margin: const EdgeInsets.only(right: 42),
               child: ElevatedButton(
                 child: const Text('Listen'),
-                onPressed: _listenLocation,
+                onPressed:
+                    _locationSubscription == null ? _listenLocation : null,
               ),
             ),
             ElevatedButton(
               child: const Text('Stop'),
-              onPressed: _stopListen,
+              onPressed: _locationSubscription != null ? _stopListen : null,
             )
           ],
         ),